add and implement order stuffs

+++ src/main/java/com/project/ecommerce/repo/ProductVariantRepository.java
@@ -3,5 +3,5 @@ package com.project.ecommerce.repo;
 import com.project.ecommerce.entitiy.ProductVariant;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProductVariantRepository extends JpaRepository<ProductVariant, Long> {
+public interface ProductVariantRepository extends JpaRepository<ProductVariant, Integer> {
 }
